Tweak isAjax and isForm checks
Summary:
We check for specific params to be passed but default the response to 'false'. If we don't set a value on the param false is true :)

This just lets the response default to null and checks that it isn't null.

Test Plan: ?__cubex_ajax__ should render an ajax request

Reviewers: brooke

Reviewed By: brooke

Differential Revision: http://phabricator.cubex.io/D24