Basic message that can be stored in session
Summary:
After seeing a lot of;

```
$msg = new \stdClass();
$msg->type = "something";
$msg->text = "something else";

Redirect::to("/")->with("msg", $msg)->now();
```

It seemed reasonable to standardise this with a very basic and transportable object.

Different session services store data in different ways, so the data may be serialized, json encoded or kept in it's original form.

For those reasons I think it's reasonable to only allow public properties and no other functionality to meet the criteria of all of the storage methods.

Test Plan: Look at it

Reviewers: brooke, rgooding, fobilow, JayFrancis

Reviewed By: fobilow

CC: fobilow, JayFrancis

Differential Revision: http://phabricator.cubex.io/D38