Basic way to look back at an extended project to load resources
Summary:
This will need to be replaced in future but needs to be backwards
compatible for now. By setting `extends = some/path` under the
`[project]` config you can set a source directory from the base of
the current project to search for any external resources.

Test Plan: Load a resource from an extended project fromit's extender.

Reviewers: brooke, fobilow, JayFrancis

Reviewed By: JayFrancis

CC: seemly

Differential Revision: http://phabricator.cubex.io/D40